requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
accelerate==0.24.1
huggingface-hub==0.19.4